jwt: fix warning from go vet
go vet (1.6.2) reports the following warning:
jwt/example_test.go:12: ExampleJWTConfig refers to unknown identifier: JWTConfig

vet ensures that in _test.go files, ExampleFoo is matched by a function or
type. In this situation, there is no type called JWTConfig. Renaming to
ExampleConfig is correct, as it creates a jwt.Config, and fixes the
warning.