oauth2: add an optional interface to add Context support to TokenSource
The main motivation for this change is that TokenSource doesn't
support contexts. We would like to access values from an http.Request
context when we are doing the request that generates a new token.

At this stage, this change is just a POC. I'd like to gather reviews
about this approach. If this approach is OK, I'll probably improve docs
and add proper testing.

Fixes golang/go#33131