oauth2: Comply with RFC8628
Fixes various small errors, possibly due to changes betweend the
device flow draft and the final RFC.

Also includes a fix to avoid being stuck in an endless poll when
the token endpoint does not return a well-formed, expected error.

The following is the example response payload in RFC8628:

      HTTP/1.1 200 OK
      Content-Type: application/json
      Cache-Control: no-store

      {
        "device_code": "GmRhmhcxhwAzkoEqiMEg_DnyEysNkuNhszIySk9eS",
        "user_code": "WDJB-MJHT",
        "verification_uri": "https://example.com/device",
        "verification_uri_complete":
            "https://example.com/device?user_code=WDJB-MJHT",
        "expires_in": 1800,
        "interval": 5
      }

Using `json:"expires_in,string"` causes issues when an unquoted int
is encountered.